requirements.txt
flask
flask-cors
requests
python-dateutil
faker
gunicorn
supabase
python-dotenv
redis

# Plane.so integration removed from this version